+++ Client/Extensions/StringExtension.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Globalization;
 
-namespace LunaClient.Utilities
+namespace LunaClient.Extensions
 {
-    public static class ExtensionMethods
+    public static class StringExtension
     {
         /// <summary>
         /// Returns only 2 decimals in case a module value has several to avoid reloading the part.
@@ -12,7 +12,7 @@ namespace LunaClient.Utilities
         /// </summary>
         public static string FormatModuleValue(this string str)
         {
-            return decimal.TryParse(str, out var decimalValue) ? 
+            return decimal.TryParse(str, out var decimalValue) ?
                 Math.Round(decimalValue, 2).ToString(CultureInfo.InvariantCulture) : str;
         }
 
@@ -23,14 +23,6 @@ namespace LunaClient.Utilities
         {
             return obj is IConvertible convertible ? convertible.ToString(CultureInfo.InvariantCulture)
                 : (obj as IFormattable)?.ToString(null, CultureInfo.InvariantCulture) ?? obj.ToString();
-        }
-
-        public static string PrintOrbitData(this Orbit orbit)
-        {
-            return $"{orbit.inclination.ToString(CultureInfo.InvariantCulture)};{orbit.eccentricity.ToString(CultureInfo.InvariantCulture)};" +
-                $"{orbit.semiMajorAxis.ToString(CultureInfo.InvariantCulture)};{orbit.LAN.ToString(CultureInfo.InvariantCulture)};" +
-                $"{orbit.argumentOfPeriapsis.ToString(CultureInfo.InvariantCulture)};{orbit.meanAnomalyAtEpoch.ToString(CultureInfo.InvariantCulture)};" +
-                $"{orbit.epoch.ToString(CultureInfo.InvariantCulture)}";
         }
     }
 }
